api_yamdb/requirements.txt
requests==2.28.2
Django==3.2.18
djangorestframework==3.14.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==5.2.2
django-filter==22.1
python-dotenv==0.21.0
gunicorn==20.0.4
psycopg2-binary==2.9.5